# Conflicts:
#	Example/Podfile.lock
#	Example/el-grocer-shopper-sdk-iOS.xcodeproj/project.pbxproj
#	el-grocer-shopper-sdk-iOS/Classes/Sources/API/AlgoliaApi.swift
#	el-grocer-shopper-sdk-iOS/Classes/Sources/Model/Entities+CRUD/Product+CRUD.swift
#	el-grocer-shopper-sdk-iOS/Classes/Sources/Modules/Dashboard/Categories/Cells/HomeCell/HomeCellViewModel.swift
#	el-grocer-shopper-sdk-iOS/Classes/Sources/Modules/Dashboard/Products/ProductsViewController.swift

+++ el-grocer-shopper-sdk-iOS/Classes/Sources/Model/Entities+CRUD/Product+CRUD.swift
@@ -531,7 +531,7 @@ extension Product {
                     product.promotionOnly = false
                 }
 
-                if product.promotionOnly.boolValue && product.promotion == 0 {
+                if product.promotionOnly.boolValue && product.promotion == 0 && onlyCurrentStoreProducts {
                     context.delete(product)
                     continue
                 }
